feat: refactor skills display into separate SkillsGrid component
